Digital cv/requirements.txt
Pillow==9.5.0
Pillow==10.2.0
streamlit==1.29.0
